Consolidate rule helpers, and keep outside rules directory for simplicity
