+++ src/rules/require-source-directory.ts
@@ -1,5 +1,6 @@
-import { buildRule, directoryExists } from './helpers';
+import { buildRule } from './build-rule';
 import { RuleName } from './types';
+import { directoryExists } from '../rule-helpers';
 
 export default buildRule({
   name: RuleName.RequireSourceDirectory,
